Implement a sample which uses the SpatialTemporalTrace class to read and display valid xml files in a folder.

+++ Multiscale/modules/verification/spatial-temporal/src/model/SpatialTemporalTrace.cpp
@@ -79,7 +79,7 @@ void SpatialTemporalTrace::initialise() {
 }
 
 void SpatialTemporalTrace::updateLastTimePointValue(TimePoint &timePoint) {
-    double timePointValue = timePoint.getValue();
+    unsigned long timePointValue = timePoint.getValue();
 
     validateTimePointValue(timePointValue);
     updateLastTimePointValue(timePoint, timePointValue);
